loader/requirements.txt
pandas>=2.2
requests>=2.32
psycopg2-binary>=2.9
python-dateutil>=2.9
